Rename clear-clipboard.js to clipboard.js and related variables
